+++ lib/titles.php
@@ -15,35 +15,32 @@ function title() {
   } elseif (is_archive()) {
     return get_the_archive_title();
   } elseif (is_search()) {
-    $title = 'Search Results ';
+
+    if (isset($_GET['proceeding_only']) && $_GET['proceeding_only'] != '') {
+      $title = '<span class="accent color">Proceeding</span> Search Results ';
+    } else {
+      $title = 'Search Results ';
+    }
+
 // for search term
     if (isset($_GET['s']) && $_GET['s'] != '') {
       $title .= 'for <span class="accent color">' . $_GET['s'] . '</span> ';
     }
-// in post type
-    if (isset($_GET['type'])) {
-      $title .= 'in';
-      $count = sizeof($_GET['type']);
-
-      foreach ($_GET['type'] as $type) {
-        $count -= 1;
-        $object = get_post_type_object($type);
-        $type = $object->labels->name;
-        $title .= ' <span class="accent color">' . ucfirst($type) . '</span>';
-
-        if ($count > 1 && sizeof($_GET['type']) > 2) {
-          $title .= ',';
-        } elseif ($count > 0 && sizeof($_GET['type']) > 2) {
-          $title .= ', and ';
-        } elseif ($count == 1) {
-          $title .= ' and ';
-        }
-      }
+// days
+    if (isset($_GET['days']) && $_GET['days'] != '') {
+      $title .= 'on <span class="accent color">'.date("F j", strtotime($_GET['days'])).'</span> ';
     }
+// sessions
+    if (isset($_GET['sessions']) && $_GET['sessions'] != '') {
+      $title .= 'during <span class="accent color">'.get_the_title($_GET['sessions']).'</span> ';
+    }
+// authors
+    if (isset($_GET['authors']) && $_GET['authors'] != '') {
+      $author = get_term_by('slug', $_GET['authors'], 'byline');
+      $title .= 'by <span class="accent color">'.$author->name.'</span> ';
+    }
+
     return __($title, 'sage');
-
-
-
 
   } elseif (is_404()) {
     return __('Not Found', 'sage');