+++ templates/content-author.php
@@ -4,7 +4,7 @@
       <?php $session = get_field('session');
       $sessionTitle = get_the_title($session); ?>
     <h2 class="entry-title"><a href="<?php the_permalink(); ?>"><?php the_title(); ?></a>
-    <br /> <small class="text-muted"><a class="accent color" href="<?php the_permalink($session); ?>"><?php echo $sessionTitle; ?></a></small></h2>
+    <br /> <small class="accent color"><?php echo $sessionTitle; ?></small></h2>
     <?php } else { ?>
     <h2 class="entry-title"><a href="<?php the_permalink(); ?>"><?php the_title(); ?></a>
     | <small class="text-muted" datetime="<?= get_post_time('c', true); ?>"><?= get_the_date(); ?></small></h2>